requirements.txt
# Core data processing
nflreadpy>=0.1.0
pandas>=2.0.0
pyarrow>=12.0.0
numpy>=1.24.0

# HTTP requests and retries
requests>=2.31.0
tenacity>=8.2.0
urllib3>=2.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0

# Data validation
pydantic>=2.0.0
jsonschema>=4.17.0

# Caching
diskcache>=5.6.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: For advanced caching
redis>=4.6.0

# Optional: For data visualization (future use)
matplotlib>=3.7.0
seaborn>=0.12.0

# API framework
fastapi>=0.104.0
uvicorn>=0.24.0
